clean up tests a bit before proceeding
