+++ datacapture/src/main/java/com/google/android/fhir/datacapture/extensions/MoreTypes.kt
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.common.datatype
+package com.google.android.fhir.datacapture.extensions
 
 import android.content.Context
 import com.google.android.fhir.datacapture.R
+import com.google.android.fhir.datacapture.fhirpath.fhirPathEngine
 import com.google.android.fhir.datacapture.format
 import com.google.android.fhir.datacapture.toLocalizedString
 import com.google.android.fhir.datacapture.views.factories.localDate
@@ -30,6 +31,7 @@ import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.DateTimeType
 import org.hl7.fhir.r4.model.DateType
 import org.hl7.fhir.r4.model.DecimalType
+import org.hl7.fhir.r4.model.Expression
 import org.hl7.fhir.r4.model.IdType
 import org.hl7.fhir.r4.model.IntegerType
 import org.hl7.fhir.r4.model.PrimitiveType
@@ -113,3 +115,19 @@ internal fun StringType.toIdType(): IdType {
 internal fun Coding.toCodeType(): CodeType {
   return CodeType(code)
 }
+
+fun Type.valueOrCalculateValue(): Type? {
+  return if (this.hasExtension()) {
+    this.extension
+      .firstOrNull { it.url == CQF_CALCULATED_EXPRESSION_URL }
+      ?.let {
+        val expression = (it.value as Expression).expression
+        fhirPathEngine.evaluate(this, expression).singleOrNull()?.let { it as Type }
+      }
+  } else {
+    this
+  }
+}
+
+internal const val CQF_CALCULATED_EXPRESSION_URL: String =
+  "http://hl7.org/fhir/StructureDefinition/cqf-calculatedValue"
