+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/factories/AutoCompleteViewHolderFactory.kt
@@ -24,7 +24,7 @@ import androidx.core.view.children
 import androidx.core.view.get
 import androidx.core.view.isEmpty
 import com.google.android.fhir.datacapture.R
-import com.google.android.fhir.datacapture.common.datatype.displayString
+import com.google.android.fhir.datacapture.extensions.displayString
 import com.google.android.fhir.datacapture.validation.Invalid
 import com.google.android.fhir.datacapture.validation.NotValidated
 import com.google.android.fhir.datacapture.validation.Valid
