+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/factories/CheckBoxGroupViewHolderFactory.kt
@@ -25,7 +25,7 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import com.google.android.fhir.datacapture.ChoiceOrientationTypes
 import com.google.android.fhir.datacapture.R
 import com.google.android.fhir.datacapture.choiceOrientation
-import com.google.android.fhir.datacapture.common.datatype.displayString
+import com.google.android.fhir.datacapture.extensions.displayString
 import com.google.android.fhir.datacapture.itemAnswerOptionImage
 import com.google.android.fhir.datacapture.optionExclusive
 import com.google.android.fhir.datacapture.validation.Invalid
