+++ datacapture/src/test/java/com/google/android/fhir/datacapture/extensions/MoreQuestionnaireItemAnswerOptionComponentsTest.kt
@@ -20,7 +20,6 @@ import android.app.Application
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
 import com.google.android.fhir.datacapture.EXTENSION_OPTION_EXCLUSIVE_URL
-import com.google.android.fhir.datacapture.common.datatype.displayString
 import com.google.android.fhir.datacapture.optionExclusive
 import com.google.common.truth.Truth.assertThat
 import java.util.Locale
