+++ datacapture/src/test/java/com/google/android/fhir/datacapture/extensions/MoreQuestionnaireResponseItemAnswerComponentsTest.kt
@@ -20,7 +20,6 @@ import android.app.Application
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
 import com.google.android.fhir.datacapture.R
-import com.google.android.fhir.datacapture.common.datatype.displayString
 import com.google.android.fhir.datacapture.format
 import com.google.android.fhir.datacapture.hasDifferentAnswerSet
 import com.google.android.fhir.datacapture.toLocalizedString
