+++ datacapture/src/test/java/com/google/android/fhir/datacapture/extensions/MoreTypesTest.kt
@@ -18,12 +18,9 @@ package com.google.android.fhir.datacapture.extensions
 
 import android.os.Build
 import ca.uhn.fhir.model.api.TemporalPrecisionEnum
-import com.google.android.fhir.datacapture.common.datatype.asStringValue
-import com.google.android.fhir.datacapture.common.datatype.toCodeType
-import com.google.android.fhir.datacapture.common.datatype.toIdType
-import com.google.android.fhir.datacapture.common.datatype.toUriType
 import com.google.common.truth.Truth.assertThat
 import java.time.Instant
+import java.time.LocalDate
 import java.time.ZoneId
 import java.util.Date
 import java.util.TimeZone
@@ -33,7 +30,10 @@ import org.hl7.fhir.r4.model.CanonicalType
 import org.hl7.fhir.r4.model.CodeType
 import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.DateTimeType
+import org.hl7.fhir.r4.model.DateType
 import org.hl7.fhir.r4.model.DecimalType
+import org.hl7.fhir.r4.model.Expression
+import org.hl7.fhir.r4.model.Extension
 import org.hl7.fhir.r4.model.IdType
 import org.hl7.fhir.r4.model.InstantType
 import org.hl7.fhir.r4.model.IntegerType
@@ -197,5 +197,30 @@ class MoreTypesTest {
   fun coding_toCodeType() {
     val code = Coding("fakeSystem", "fakeCode", "fakeDisplay").toCodeType()
     assertThat(code.equalsDeep(CodeType("fakeCode"))).isTrue()
+  }
+
+  @Test
+  fun `should return calculated value for cqf expression`() {
+    val today = LocalDate.now().toString()
+    val type =
+      DateType().apply {
+        extension =
+          listOf(
+            Extension(
+              CQF_CALCULATED_EXPRESSION_URL,
+              Expression().apply {
+                language = "text/fhirpath"
+                expression = "today()"
+              }
+            )
+          )
+      }
+    assertThat((type.valueOrCalculateValue() as? DateType)?.valueAsString).isEqualTo(today)
+  }
+
+  @Test
+  fun `should return entered value when no cqf expression is defined`() {
+    val type = IntegerType().apply { value = 500 }
+    assertThat((type.valueOrCalculateValue() as? IntegerType)?.value).isEqualTo(500)
   }
 }
