+++ datacapture/src/test/java/com/google/android/fhir/datacapture/fhirpath/ExpressionEvaluatorTest.kt
@@ -18,7 +18,7 @@ package com.google.android.fhir.datacapture.fhirpath
 
 import com.google.android.fhir.datacapture.EXTENSION_CALCULATED_EXPRESSION_URL
 import com.google.android.fhir.datacapture.EXTENSION_VARIABLE_URL
-import com.google.android.fhir.datacapture.common.datatype.asStringValue
+import com.google.android.fhir.datacapture.extensions.asStringValue
 import com.google.android.fhir.datacapture.extensions.variableExpressions
 import com.google.android.fhir.datacapture.fhirpath.ExpressionEvaluator.detectExpressionCyclicDependency
 import com.google.android.fhir.datacapture.fhirpath.ExpressionEvaluator.evaluateCalculatedExpressions
